chat panel component
